refactor(stats): rename app
