+++ config/urls.py
@@ -11,11 +11,11 @@ from lacommunaute.forum_conversation import urls as forum_conversation_extension
 from lacommunaute.forum_conversation.forum_polls import urls as forum_polls_extension_urls
 from lacommunaute.forum_member import urls as forum_member_urls
 from lacommunaute.forum_moderation import urls as forum_moderation_urls
-from lacommunaute.forum_stats import urls as forum_stats_urls
 from lacommunaute.forum_upvote import urls as forum_upvote_urls
 from lacommunaute.inclusion_connect import urls as inclusion_connect_urls
 from lacommunaute.pages import urls as pages_urls
 from lacommunaute.search import urls as search_urls
+from lacommunaute.stats import urls as stats_urls
 from lacommunaute.surveys import urls as surveys_urls
 
 
@@ -38,7 +38,7 @@ urlpatterns = [
     path("", include(forum_moderation_urls)),
     path("calendar/", include(event_urls)),
     path("surveys/", include(surveys_urls)),
-    path("statistiques/", include(forum_stats_urls)),
+    path("statistiques/", include(stats_urls)),
     # machina legacy
     path("", include(conversation_urlpatterns_factory.urlpatterns)),
     path("moderation/", include(moderation_urlpatterns_factory.urlpatterns)),
