+++ lacommunaute/stats/apps.py
@@ -2,7 +2,7 @@ from django.apps import AppConfig
 
 
 class ForumStatsAppConfig(AppConfig):
-    label = "forum_stats"
-    name = "lacommunaute.forum_stats"
+    label = "stats"
+    name = "lacommunaute.stats"
     verbose_name = "Forum Stats"
     verbose_name_plural = "Forum Stats"
