+++ lacommunaute/stats/management/commands/collect_django_stats.py
