+++ lacommunaute/templates/stats/base_detail_stats.html
@@ -10,7 +10,7 @@
             <ol class="breadcrumb">
                 <li class="breadcrumb-item">{% trans "Back to" %}</li>
                 <li class="breadcrumb-item">
-                    <a href="{% url 'forum_stats:statistiques' %}">Statistiques</a>
+                    <a href="{% url 'stats:statistiques' %}">Statistiques</a>
                 </li>
             </ol>
         </nav>
